Struct limine::file::File

#[repr(C)]
pub struct File { /* private fields */ }

A file loaded by the bootloader. Returned from KernelFileRequest and ModuleRequest.

Implementations

impl File

pub fn revision(&self) -> u64

Get the revision of the file. Currently, this is always 0.

pub fn addr(&self) -> *mut u8

The base address of the file. Note that this is not necessarily a pointer to executable code. It simply points to the raw file.

pub fn size(&self) -> u64

The size of the file, in bytes.

pub fn path(&self) -> &[u8]

The path of the file. This is the path that was passed to the bootloader in either the configuration file or the internal_modules field of the ModuleRequest.

It is returned as a raw byte slice, and the encoding is unspecified.

pub fn cmdline(&self) -> &[u8]

The command line of the file. This is the command line that was passed to the bootloader in either the configuration file or the internal_modules field of the ModuleRequest.

It is returned as a raw byte slice, and the encoding is unspecified.

pub fn media_type(&self) -> MediaType

The media type of the file. See MediaType for more information.

pub fn tftp_ip(&self) -> Option<NonZeroU32>

The IP address of the TFTP server, if the file was loaded from a TFTP.

pub fn tftp_port(&self) -> Option<NonZeroU32>

The port of the TFTP server, if the file was loaded from a TFTP.

pub fn partition_idx(&self) -> Option<NonZeroU32>

The partition index of the file, if the file was loaded from a partition.

pub fn mbr_disk_id(&self) -> Option<NonZeroU32>

The MBR disk ID of the file, if the file was loaded from an MBR disk.

pub fn gpt_disk_id(&self) -> Option<Uuid>

The GPT disk UUID of the file, if the file was loaded from a GPT disk.

pub fn gpt_partition_id(&self) -> Option<Uuid>

The GPT partition UUID of the file, if the file was loaded from a GPT partition.

pub fn partition_uuid(&self) -> Option<Uuid>

The partition UUID of the file, if the file was loaded from a partition with a UUID.