limine::framebuffer

Struct Framebuffer

pub struct Framebuffer<'a> { /* private fields */ }

A pointer to a framebuffer.

Why is this a wrapper type?

Two revisions currently exist of the framebuffer type. However, the type itself has no revision field. In order to keep this type safe, we wrap the pointer with its associated revision taken from the response.

Implementations

impl<'a> Framebuffer<'a>

pub fn addr(&self) -> *mut u8

The address of the framebuffer. Note that no synchronization is done on this pointer, so you yourself must synchronize all access. In addition, the pointer may point to uninitialized bytes at boot, so dereferencing it will cause UB.

pub fn width(&self) -> u64

The width of the framebuffer in its current mode.

pub fn height(&self) -> u64

The height of the framebuffer in its current mode.

pub fn pitch(&self) -> u64

The pitch (distance between rows, in bytes) of the framebuffer in the current mode. This is not always the same as (width * bpp) / 8, as padding bytes may be added to achieve a particular alignment.

pub fn bpp(&self) -> u16

The number of bits (not bytes) per pixel of the framebuffer in the current mode.

pub fn memory_model(&self) -> MemoryModel

The memory model of the framebuffer in the current mode. See MemoryModel for specific values.

pub fn red_mask_size(&self) -> u8

The size of the red mask, in bits. This part of the mask can be applied with red_value & ((1 << red_mask_size) - 1).

pub fn red_mask_shift(&self) -> u8

The number of bits to shift the red mask to the left. This part of the mask can be applied with red_value << red_mask_shift.

pub fn green_mask_size(&self) -> u8

The size of the green mask, in bits. This part of the mask can be applied with green_value & ((1 << green_mask_size) - 1).

pub fn green_mask_shift(&self) -> u8

The number of bits to shift the green mask to the left. This part of the mask can be applied with green_value << green_mask_shift.

pub fn blue_mask_size(&self) -> u8

The size of the blue mask, in bits. This part of the mask can be applied with blue_value & ((1 << blue_mask_size) - 1).

pub fn blue_mask_shift(&self) -> u8

The number of bits to shift the blue mask to the left. This part of the mask can be applied with blue_value << blue_mask_shift.

pub fn edid(&self) -> &[u8]

The raw EDID bytes of the display attached to this framebuffer.

pub fn modes(&self) -> Option<&[&VideoMode]>

The video modes supported on this framebuffer. Only available on revision 1 and above.