limine::framebuffer

Struct VideoMode

#[repr(C)]
pub struct VideoMode {
    pub pitch: u64,
    pub width: u64,
    pub height: u64,
    pub bpp: u16,
    pub memory_model: MemoryModel,
    pub red_mask_size: u8,
    pub red_mask_shift: u8,
    pub green_mask_size: u8,
    pub green_mask_shift: u8,
    pub blue_mask_size: u8,
    pub blue_mask_shift: u8,
}

A mode supported by the current framebuffer.

Fields
pitch: u64

The pitch (distance between rows, in bytes). This is not always the same as (width * bpp) / 8, as padding bytes may be added to achieve a particular alignment.

width: u64

The width of the framebuffer, in pixels.

height: u64

The height of the framebuffer, in pixels.

bpp: u16

The number of bits (not bytes) per single pixel in the framebuffer.

memory_model: MemoryModel

The memory model used by this video mode. See MemoryModel for specific values.

red_mask_size: u8

The size of the red mask, in bits. This part of the mask can be applied with red_value & ((1 << red_mask_size) - 1).

red_mask_shift: u8

The number of bits to shift the red mask to the left. This part of the mask can be applied with red_value << red_mask_shift.

green_mask_size: u8

The size of the green mask, in bits. This part of the mask can be applied with green_value & ((1 << green_mask_size) - 1).

green_mask_shift: u8

The number of bits to shift the green mask to the left. This part of the mask can be applied with green_value << green_mask_shift.

blue_mask_size: u8

The size of the blue mask, in bits. This part of the mask can be applied with blue_value & ((1 << blue_mask_size) - 1).

blue_mask_shift: u8

The number of bits to shift the blue mask to the left. This part of the mask can be applied with blue_value << blue_mask_shift.