limine::modules

Struct InternalModule

#[repr(C)]
pub struct InternalModule { /* private fields */ }

An internal module that the kernel requests from the bootloader. Only available with request revision 1 and greater.

Implementations

impl InternalModule

pub const fn new() -> Self

Create a new internal module with no path, no cmdline and no flags.

pub const fn with_path(self, path: &'static CStr) -> Self

Set the path of the internal module. This function returns the new value.

Parameters

path: The new value of the field.

pub fn set_path(&mut self, path: &'static CStr)

Set the path of the internal module. This function operates in place.

Parameters

path: The new value of the field.

pub const fn with_cmdline(self, cmdline: &'static CStr) -> Self

Set the command-line for the module. This function returns the new value.

Parameters

cmdline: The new value of the field.

pub fn set_cmdline(&mut self, cmdline: &'static CStr)

Set the command-line for the module. This function operates in place.

Parameters

cmdline: The new value of the field.

pub const fn with_flags(self, flags: ModuleFlags) -> Self

Set the flags for the module. This function returns the new value.

Parameters

flags: The new value of the field.

pub fn set_flags(&mut self, flags: ModuleFlags)

Set the flags for the module. This function operates in place.

Parameters

flags: The new value of the field.

pub fn path(&self) -> &[u8]

Returns the module’s path as a byte slice with unspecified encoding.

pub fn cmdline(&self) -> &[u8]

Returns the module’s command-line as a byte slice with unspecified encoding.

pub fn flags(&self) -> ModuleFlags

Returns the module’s flags.