limine::request

Struct DeviceTreeBlobRequest

#[repr(C)]
pub struct DeviceTreeBlobRequest { /* private fields */ }

Request the address of the device-tree blob, if one is present.

Usage

static BASE_REVISION: BaseRevision = BaseRevision::new();

// Request the device-tree blob address
static DEVICE_TREE_BLOB_REQUEST: DeviceTreeBlobRequest = DeviceTreeBlobRequest::new();

// ...later, in our code
DEVICE_TREE_BLOB_REQUEST.get_response() // ...

Implementations

impl DeviceTreeBlobRequest

pub const fn new() -> Self

Create a new request with the latest revision.

pub const fn with_revision(revision: u64) -> Self

Create a new request with the given revision.

pub fn id(&self) -> &[u64; 4]

Get the ID of this request. This includes the magic number and the request-specific ID.

pub fn revision(&self) -> u64

Get the revision of this request.

pub fn get_response(&self) -> Option<&DeviceTreeBlobResponse>

Get the response to this request, if it has been set.

pub fn get_response_mut(&mut self) -> Option<&mut DeviceTreeBlobResponse>

Get a mutable reference to the response to this request, if it has been set.

Note that this method takes a mutable reference, so the request will have to be wrapped in a mutex or similar in order to use it.