Struct limine::request::KernelAddressRequest

#[repr(C)]
pub struct KernelAddressRequest { /* private fields */ }

Request the base address of the kernel code, in virtual and physical space.

Usage

static BASE_REVISION: BaseRevision = BaseRevision::new();

// Request the kernel address
static KERNEL_ADDRESS_REQUEST: KernelAddressRequest = KernelAddressRequest::new();

// ...later, in our code
KERNEL_ADDRESS_REQUEST.get_response() // ...

Implementations

impl KernelAddressRequest

pub const fn new() -> Self

Create a new request with the latest revision.

pub const fn with_revision(revision: u64) -> Self

Create a new request with the given revision.

pub fn id(&self) -> &[u64; 4]

Get the ID of this request. This includes the magic number and the request-specific ID.

pub fn revision(&self) -> u64

Get the revision of this request.

pub fn get_response(&self) -> Option<&KernelAddressResponse>

Get the response to this request, if it has been set.

pub fn get_response_mut(&mut self) -> Option<&mut KernelAddressResponse>

Get a mutable reference to the response to this request, if it has been set.

Note that this method takes a mutable reference, so the request will have to be wrapped in a mutex or similar in order to use it.