Struct limine::request::PagingModeRequest

#[repr(C)]
pub struct PagingModeRequest { /* private fields */ }

Request certain platform-dependent paging modes and flags to be set.

Usage

static BASE_REVISION: BaseRevision = BaseRevision::new();

// Request a paging mode
#[cfg(any(target_arch = "x86_64", target_arch = "aarch64"))] // x86_64 and AArch64 share the same modes
static PAGING_MODE_REQUEST: PagingModeRequest = PagingModeRequest::new().with_mode(paging::Mode::FOUR_LEVEL);

#[cfg(target_arch = "riscv64")] // RISC-V has different modes
static PAGING_MODE_REQUEST: PagingModeRequest = PagingModeRequest::new().with_mode(paging::Mode::SV48);

// ...later, in our code
PAGING_MODE_REQUEST.get_response() // ...

Implementations

impl PagingModeRequest

pub const fn new() -> Self

Create a new request with the latest revision.

pub const fn with_revision(revision: u64) -> Self

Create a new request with the given revision.

pub fn id(&self) -> &[u64; 4]

Get the ID of this request. This includes the magic number and the request-specific ID.

pub fn revision(&self) -> u64

Get the revision of this request.

pub fn get_response(&self) -> Option<&PagingModeResponse>

Get the response to this request, if it has been set.

pub fn get_response_mut(&mut self) -> Option<&mut PagingModeResponse>

Get a mutable reference to the response to this request, if it has been set.

Note that this method takes a mutable reference, so the request will have to be wrapped in a mutex or similar in order to use it.

pub fn set_mode(&mut self, mode: impl Into<Mode>)

Set the requested paging mode. See Mode for more information. This function operates in place.

Parameters

mode: The new value for the field.

pub const fn with_mode(self, mode: Mode) -> Self

Set the requested paging mode. See Mode for more information. This function returns the new value.

Parameters

mode: The new value for the field.

pub fn set_flags(&mut self, flags: impl Into<Flags>)

Set the requested paging flags. See Flags for more information. This function operates in place.

Parameters

flags: The new value for the field.

pub const fn with_flags(self, flags: Flags) -> Self

Set the requested paging flags. See Flags for more information. This function returns the new value.

Parameters

flags: The new value for the field.

pub fn mode(&self) -> Mode

Get the requested paging mode. See Mode for more information.

pub fn flags(&self) -> Flags

Get the requested paging flags. See Flags for more information.