limine::response

Struct HhdmResponse

#[repr(C)]
pub struct HhdmResponse { /* private fields */ }

A response to a higher-half direct map request.

Implementations

impl HhdmResponse

pub fn revision(&self) -> u64

Returns the revision of the response.

pub fn offset(&self) -> u64

Returns the offset of the higher-half direct map. This can be used to convert physical addresses to virtual addresses, and the same in reverse, as long as the bootloader’s page tables are still in use.

Examples

Convert physical to virtual:

let virt_addr = phys_addr + offset;

Convert virtual† to physical:

let phys_addr = virt_addr - offset;

† Note that this only works if the virtual address is in the higher-half direct map already. This is true of any virtual address returned in any response. However, this is not true of addresses within executable code. To convert an address within executable code, you must use the kernel address request.