Struct limine::response::MemoryMapResponse

#[repr(C)]
pub struct MemoryMapResponse { /* private fields */ }

A response to a memory map request.

Implementations

impl MemoryMapResponse

pub fn revision(&self) -> u64

Returns the revision of the response.

pub fn entries(&self) -> &[&Entry]

Returns a slice of found memory map entries. See Entry for more information.

pub fn entries_mut(&mut self) -> &mut [&mut Entry]

Returns a mutable slice of found memory map entries. See Entry for more information. Note that this function takes &mut self, so the response will likely need to be wrapped in a Mutex or similar.