limine::response

Struct SmpResponse

#[repr(C)]
pub struct SmpResponse { /* private fields */ }

A response to a smp request. This response contains information about the boot processor and all other processors.

Implementations

impl SmpResponse

pub fn revision(&self) -> u64

Returns the revision of the response.

pub fn flags(&self) -> ResponseFlags

Returns the flags that were enabled by the bootloader. See ResponseFlags for more information.

pub fn bsp_lapic_id(&self) -> u32

Returns the local APIC ID of the boot processor. This is only available on x86_64.

pub fn cpus(&self) -> &[&Cpu]

Returns a slice of found CPUs. See [Cpu] for more information.

pub fn cpus_mut(&mut self) -> &mut [&mut Cpu]

Returns a mutable slice of found CPUs. See [Cpu] for more information. Note that this function takes &mut self, so the response will likely need to be wrapped in a Mutex or similar. It is provided so that the extra field on each CPU can be set.