Struct limine::smp::Cpu

#[repr(C)]
pub struct Cpu {
    pub id: u32,
    pub lapic_id: u32,
    pub goto_address: GotoAddress,
    pub extra: u64,
    /* private fields */
}

A CPU entry in the SMP request.

Fields
id: u32

The ACPI processor ID, according to the ACPI MADT.

lapic_id: u32

The APIC ID, according to the ACPI MADT.

goto_address: GotoAddress

The address to jump to. Writing to this field will cause the core to jump to the given function. The function will receive a pointer to this structure, and it will have its own 64KiB (or requested-size) stack.

extra: u64

Free for use by the kernel.