Struct limine::LimineFile

pub struct LimineFile {

    pub revision: u64,
    pub base: LiminePtr<u8>,
    pub length: u64,
    pub path: LiminePtr<c_char>,
    pub cmdline: LiminePtr<c_char>,
    pub media_type: u64,
    pub unused: u32,
    pub tftp_ip: u32,
    pub tftp_port: u32,
    pub partition_index: u32,
    pub mbr_disk_id: u32,
    pub gpt_disk_uuid: LimineUuid,
    pub gpt_part_uuid: LimineUuid,
    pub part_uuid: LimineUuid,

}

Fields
revision: u64

Revision of this structure.

base: LiminePtr<u8>

The address of the file.

length: u64

The size of the file.

path: LiminePtr<c_char>

The path of the file within the volume, with a leading slash.

cmdline: LiminePtr<c_char>

A command line associated with the file.

media_type: u64

Type of media file resides on.

unused: u32tftp_ip: u32

If non-0, this is the IP of the TFTP server the file was loaded from.

tftp_port: u32

Likewise, but port.

partition_index: u32

1-based partition index of the volume from which the file was loaded. If 0, it means invalid or unpartitioned.

mbr_disk_id: u32

If non-0, this is the ID of the disk the file was loaded from as reported in its MBR.

gpt_disk_uuid: LimineUuid

If non-0, this is the UUID of the disk the file was loaded from as reported in its GPT.

gpt_part_uuid: LimineUuid

If non-0, this is the UUID of the partition the file was loaded from as reported in the GPT.

part_uuid: LimineUuid

If non-0, this is the UUID of the filesystem of the partition the file was loaded from.