Crate limine

Rust crate for parsing the limine boot protocol structures.

Resources

Limine Boot Protocol Specification

Structs

BootInfoRequest

BootInfoResponse

BootTimeRequest

BootTimeResponse

DtbRequest

DtbResponse

Notes

EfiSystemTableRequest

EfiSystemTableResponse

EntryPoint

EntryPointRequest

EntryPointResponse

File

Framebuffer

FramebufferRequest

FramebufferResponse

HhdmRequest

HhdmResponse

KernelAddressRequest

KernelAddressResponse

KernelFileRequest

KernelFileResponse

Level5PagingRequestDeprecated

The presence of this request will prompt the bootloader to turn on x86_64 5-level paging. It will not be turned on if this request is not present. If the response pointer is unchanged, 5-level paging is engaged.

Level5PagingResponseDeprecated

MemmapEntry

MemmapRequest

MemmapResponse

ModuleRequest

ModuleResponse

NonNullPtr

[NonNull] with the dereference traits implemented.

PagingModeRequest

PagingModeResponse

Ptr

RsdpRequest

RsdpResponse

SmbiosRequest

SmbiosResponse

SmpInfo

SmpRequest

The presence of this request will prompt the bootloader to bootstrap the secondary processors. This will not be done if this request is not present.

SmpResponse

StackSizeRequest

StackSizeResponse

TerminalDeprecated

TerminalRequestDeprecated

Omitting this request will cause the bootloader to not initialise the terminal service.

TerminalResponseDeprecated

Uuid

Enums

MemoryMapEntryType

PagingMode