Crate limine

Limine Boot Protocol

This modules implements the kernel-facing interface provided by the Limine Boot Protocol, a modern, minimal boot protocol for 64-bit higher-half kernels.

Features

The protocol is built on a mechanism of requests and responses, collectively called “features”. The kernel requests various features from the bootloader by declaring requests in the kernel binary. Each request contains a 256-bit identifier used to locate it within the kernel binary, along with a revision number and an empty response pointer field which is filled by the bootloader if the requested feature is supported.

static BOOTLOADER_INFO_REQUEST: BootloaderInfoRequest = BootloaderInfoRequest::new();

fn main() {
    if let Some(response) = BOOTLOADER_INFO_REQUEST.response() {
        println!("{} {}", response.name(), response.version());
    } else {
        println!("the bootloader info feature is not supported by the bootloader!");
    }
}

Macros

declare_feature

Structs

BootTime

Reports the system time on boot

BootTimeRequest

Request for the BootTime feature

BootloaderInfo

BootloaderInfoRequest

Request for the BootloaderInfo feature

Dtb

Device Tree Blob

DtbRequest

Request for the Dtb feature

EfiSystemTable

EfiSystemTableRequest

Request for the EfiSystemTable feature

EntryPoint

EntryPointRequest

Request for the EntryPoint feature

File

FiveLevelPaging

FiveLevelPagingRequest

Request for the FiveLevelPaging feature

Framebuffer

FramebufferRequest

Request for the Framebuffers feature

Framebuffers

Hhdm

HhdmRequest

Request for the Hhdm feature

KernelAddress

Reports the virtual and physical base addresses of the kernel

KernelAddressRequest

Request for the KernelAddress feature

KernelFile

KernelFileRequest

Request for the KernelFile feature

LiminePtr

A pointer to a resource provided by the bootloader

MemoryMap

MemoryMapEntry

MemoryMapRequest

Request for the MemoryMap feature

Modules

ModulesRequest

Request for the Modules feature

PixelFormat

RequestId

Request Identifier

Rsdp

RsdpRequest

Request for the Rsdp feature

Smbios

SmbiosRequest

Request for the Smbios feature

Smp

SmpFlags

SmpInfo

SmpRequest

Request for the Smp feature

SmpRequestFlags

SmpStartupError

StackSize

StackSizeRequest

Request for the StackSize feature

Uuid

Enums

MemoryKind

Traits

LimineRequest

Type Definitions

EntryPointFn