Crate limine

Rust crate for parsing the limine boot protocol structures.

Resources

Limine Boot Protocol Specification

Structs

Limine5LevelPagingRequest

The presence of this request will prompt the bootloader to turn on x86_64 5-level paging. It will not be turned on if this request is not present. If the response pointer is unchanged, 5-level paging is engaged.

Limine5LevelPagingResponse

LimineBootInfoRequest

LimineBootInfoResponse

LimineBootTimeRequest

LimineBootTimeResponse

LimineDtbRequest

LimineDtbResponse

Notes

LimineEfiSystemTableRequest

LimineEfiSystemTableResponse

LimineEntryPoint

LimineEntryPointRequest

LimineEntryPointResponse

LimineFile

LimineFramebuffer

LimineFramebufferRequest

LimineFramebufferResponse

LimineHhdmRequest

LimineHhdmResponse

LimineKernelAddressRequest

LimineKernelAddressResponse

LimineKernelFileRequest

LimineKernelFileResponse

LimineMemmapEntry

LimineMemmapRequest

LimineMemmapResponse

LimineModuleRequest

LimineModuleResponse

LiminePtr

LimineRsdpRequest

LimineRsdpResponse

LimineSmbiosRequest

LimineSmbiosResponse

LimineSmpInfo

LimineSmpRequest

The presence of this request will prompt the bootloader to bootstrap the secondary processors. This will not be done if this request is not present.

LimineSmpResponse

LimineStackSizeRequest

LimineStackSizeResponse

LimineTerminal

LimineTerminalRequest

Omitting this request will cause the bootloader to not initialise the terminal service.

LimineTerminalResponse

LimineUuid

NonNullPtr

[NonNull] with the dereference traits implemented.

Enums

LimineMemoryMapEntryType