limine

Module request

Request structures

Structs

BootTimeRequest

Request the boot time in seconds

BootloaderInfoRequest

Request the name and version of the loading bootloader.

DeviceTreeBlobRequest

Request the address of the device-tree blob, if one is present.

EfiMemoryMapRequest

Request the address of the UEFI memory map.

EfiSystemTableRequest

Request the address of the UEFI system table.

EntryPointRequest

Requests limine to use a specific function as the kernel entry point, instead of the one specified in the ELF.

FiveLevelPagingRequestDeprecated

This request is deprecated and was removed from the reference implementation. Use PagingModeRequest instead.

FramebufferRequest

Request a framebuffer for graphics output.

HhdmRequest

Request information about the higher-half direct mapping.

KernelAddressRequest

Request the base address of the kernel code, in virtual and physical space.

KernelFileRequest

Request information about the loaded kernel file. See File for more information.

MemoryMapRequest

Request limine’s memory map. This may or may not be the same as the one given by the BIOS/UEFI firmware. Entries are guaranteed to be in order of their base address. Usable and bootloader-reclaimable memory regions will never overlap, and will always be 4096-byte aligned. Other region types have no such guarantees. Some holes may exist. Memory between 0x0-0x1000 is never marked as usable.

ModuleRequest

Request information about the loaded modules. See File for more information.

PagingModeRequest

Request certain platform-dependent paging modes and flags to be set.

RsdpRequest

Request the RSDP address.

SmbiosRequest

Request the address of the SMBIOS table.

SmpRequest

Request the start of all other cores on the system, if they exist. Without this request, non-bootstrap cores will be ignored.

StackSizeRequest

Request a differently-sized stack.