Module limine::response

Responses to requests.

Structs

BootTimeResponse

A response to a boot time request.

BootloaderInfoResponse

A response to a bootloader info request.

DeviceTreeBlobResponse

A response to a device tree blob request.

EfiMemoryMapResponse

A response to a memory map request.

EfiSystemTableResponse

A response to a system table request.

EntryPointResponse

A response to a kernel file request.

FiveLevelPagingResponse

This request is deprecated and was removed from the reference implementation. Use PagingModeRequest instead.

FramebufferResponse

A response to a framebuffer request.

HhdmResponse

A response to a higher-half direct map request.

KernelAddressResponse

A response to a kernel address request.

KernelFileResponse

A response to a kernel file request.

MemoryMapResponse

A response to a memory map request.

ModuleResponse

A response to a module request.

PagingModeResponse

A response to a paging mode request.

RsdpResponse

A response to a rsdp request.

SmbiosResponse

A response to a smbios request.

SmpResponse

A response to a smp request. This response contains information about the boot processor and all other processors.

StackSizeResponse

A response to a stack size request. This response has no fields. If it is provided, the bootloader complied with the request.